Fix tests

+++ users/tasks.py
@@ -1,9 +1,12 @@
 import pydoc
+import logging
 
 from common.models import ErrorQueue
-from users.models import send_email_on_signup
+from users.models import send_email_on_signup, MflUser
 from celery.task.schedules import crontab
 from celery.decorators import periodic_task
+
+LOGGER = logging.getLogger(__name__)
 
 
 @periodic_task(
@@ -16,17 +19,18 @@ def resend_user_signup_emails():
     for obj in objects_with_errors:
         obj_path = "{0}.models.{1}".format(obj.app_label, obj.model_name)
         model = pydoc.locate(obj_path)
-        instance = model.objects.get(id=obj.object_pk)
         try:
+            instance = model.objects.get(id=obj.object_pk)
             user_id = instance.id
             email = instance.email
             first_name = instance.first_name
             employee_number = instance.employee_number
-            send_email_on_signup(
+            mail_sent = send_email_on_signup(
                 user_id, email, first_name, employee_number)
-            obj.delete()
-        except instance.DoesNotExist:
-            pass
-        except:
-            obj.retries = obj.retries + 1
-            obj.save()
+            if mail_sent:
+                obj.delete()
+            else:
+                obj.retries = obj.retries + 1
+                obj.save()
+        except MflUser.DoesNotExist:
+            LOGGER.info("The user has deleted")
